+++ Common/src/main/java/fuzs/eternalnether/common/client/renderer/entity/PiglinPrisonerRenderer.java
@@ -3,13 +3,11 @@ package fuzs.eternalnether.common.client.renderer.entity;
 import fuzs.eternalnether.common.EternalNether;
 import fuzs.eternalnether.common.client.model.geom.ModModelLayers;
 import net.minecraft.client.renderer.entity.EntityRendererProvider;
-import net.minecraft.client.renderer.entity.PiglinRenderer;
 import net.minecraft.client.renderer.entity.state.PiglinRenderState;
 import net.minecraft.resources.Identifier;
 
-public class PiglinPrisonerRenderer extends PiglinRenderer {
-    private static final Identifier TEXTURE_LOCATION = EternalNether.id(
-            "textures/entity/piglin/piglin_prisoner.png");
+public class PiglinPrisonerRenderer extends AbstractPiglinRenderer {
+    private static final Identifier TEXTURE_LOCATION = EternalNether.id("textures/entity/piglin/piglin_prisoner.png");
 
     public PiglinPrisonerRenderer(EntityRendererProvider.Context context) {
         super(context,
@@ -20,7 +18,7 @@ public class PiglinPrisonerRenderer extends PiglinRenderer {
     }
 
     @Override
-    public Identifier getTextureLocation(PiglinRenderState renderState) {
+    public Identifier getTextureLocation(PiglinRenderState state) {
         return TEXTURE_LOCATION;
     }
 }