+++ Common/src/main/java/fuzs/eternalnether/common/world/entity/monster/piglin/AgeablePiglin.java
@@ -1,0 +1,215 @@
+package fuzs.eternalnether.common.world.entity.monster.piglin;
+
+import net.minecraft.core.BlockPos;
+import net.minecraft.network.syncher.EntityDataAccessor;
+import net.minecraft.network.syncher.EntityDataSerializers;
+import net.minecraft.network.syncher.SynchedEntityData;
+import net.minecraft.resources.Identifier;
+import net.minecraft.sounds.SoundEvent;
+import net.minecraft.sounds.SoundEvents;
+import net.minecraft.util.RandomSource;
+import net.minecraft.world.DifficultyInstance;
+import net.minecraft.world.damagesource.DamageSource;
+import net.minecraft.world.entity.*;
+import net.minecraft.world.entity.ai.attributes.AttributeInstance;
+import net.minecraft.world.entity.ai.attributes.AttributeModifier;
+import net.minecraft.world.entity.ai.attributes.Attributes;
+import net.minecraft.world.entity.ai.goal.*;
+import net.minecraft.world.entity.ai.goal.target.HurtByTargetGoal;
+import net.minecraft.world.entity.ai.goal.target.NearestAttackableTargetGoal;
+import net.minecraft.world.entity.boss.wither.WitherBoss;
+import net.minecraft.world.entity.monster.piglin.AbstractPiglin;
+import net.minecraft.world.entity.monster.piglin.Piglin;
+import net.minecraft.world.entity.monster.piglin.PiglinAi;
+import net.minecraft.world.entity.monster.piglin.PiglinArmPose;
+import net.minecraft.world.entity.monster.skeleton.WitherSkeleton;
+import net.minecraft.world.entity.player.Player;
+import net.minecraft.world.level.Level;
+import net.minecraft.world.level.LevelAccessor;
+import net.minecraft.world.level.ServerLevelAccessor;
+import net.minecraft.world.level.block.Blocks;
+import net.minecraft.world.level.block.state.BlockState;
+import net.minecraft.world.level.storage.ValueInput;
+import net.minecraft.world.level.storage.ValueOutput;
+import org.jspecify.annotations.Nullable;
+
+/**
+ * A {@link Goal} based implementation of {@link AbstractPiglin}.
+ * <p>
+ * Additionally implements baby components from {@link Piglin}.
+ */
+public abstract class AgeablePiglin extends AbstractPiglin {
+    /**
+     * @see Piglin#DATA_BABY_ID
+     */
+    private static final EntityDataAccessor<Boolean> DATA_BABY_ID = SynchedEntityData.defineId(AgeablePiglin.class,
+            EntityDataSerializers.BOOLEAN);
+    /**
+     * @see Piglin#BABY_DIMENSIONS
+     */
+    private static final EntityDimensions BABY_DIMENSIONS = EntityType.PIGLIN.getDimensions()
+            .scale(0.5F)
+            .withEyeHeight(0.97F);
+    /**
+     * @see Piglin#SPEED_MODIFIER_BABY_ID
+     */
+    private static final Identifier SPEED_MODIFIER_BABY_ID = Identifier.withDefaultNamespace("baby");
+    /**
+     * @see Piglin#SPEED_MODIFIER_BABY
+     */
+    private static final AttributeModifier SPEED_MODIFIER_BABY = new AttributeModifier(SPEED_MODIFIER_BABY_ID,
+            0.2F,
+            AttributeModifier.Operation.ADD_MULTIPLIED_BASE);
+
+    public AgeablePiglin(EntityType<? extends AbstractPiglin> entityType, Level level) {
+        super(entityType, level);
+    }
+
+    /**
+     * @see Piglin#checkPiglinSpawnRules(EntityType, LevelAccessor, EntitySpawnReason, BlockPos, RandomSource)
+     */
+    public static boolean checkPiglinSpawnRules(EntityType<? extends AbstractPiglin> piglin, LevelAccessor level, EntitySpawnReason entitySpawnReason, BlockPos pos, RandomSource random) {
+        return !level.getBlockState(pos.below()).is(Blocks.NETHER_WART_BLOCK);
+    }
+
+    @Override
+    protected void registerGoals() {
+        super.registerGoals();
+        this.goalSelector.addGoal(1, new FloatGoal(this));
+        this.goalSelector.addGoal(5, new AvoidEntityGoal<>(this, Mob.class, 8.0F, 1.0, 1.2, PiglinAi::isZombified));
+        this.goalSelector.addGoal(7, new RandomStrollGoal(this, 0.6));
+        this.goalSelector.addGoal(9, new LookAtPlayerGoal(this, Player.class, 15.0F, 1.0F));
+        this.goalSelector.addGoal(10, new RandomLookAroundGoal(this));
+        this.targetSelector.addGoal(1, new HurtByTargetGoal(this));
+        this.targetSelector.addGoal(2, new NearestAttackableTargetGoal<>(this, Player.class, true));
+        this.targetSelector.addGoal(3, new NearestAttackableTargetGoal<>(this, WitherSkeleton.class, true));
+        this.targetSelector.addGoal(3, new NearestAttackableTargetGoal<>(this, WitherBoss.class, true));
+    }
+
+    @Override
+    protected void defineSynchedData(SynchedEntityData.Builder entityData) {
+        super.defineSynchedData(entityData);
+        entityData.define(DATA_BABY_ID, false);
+    }
+
+    /**
+     * @see Piglin#onSyncedDataUpdated(EntityDataAccessor)
+     */
+    @Override
+    public void onSyncedDataUpdated(EntityDataAccessor<?> accessor) {
+        super.onSyncedDataUpdated(accessor);
+        if (DATA_BABY_ID.equals(accessor)) {
+            this.refreshDimensions();
+        }
+    }
+
+    @Override
+    protected void addAdditionalSaveData(ValueOutput output) {
+        super.addAdditionalSaveData(output);
+        output.putBoolean("IsBaby", this.isBaby());
+    }
+
+    @Override
+    protected void readAdditionalSaveData(ValueInput input) {
+        super.readAdditionalSaveData(input);
+        this.setBaby(input.getBooleanOr("IsBaby", false));
+    }
+
+    /**
+     * @see Piglin#getDefaultDimensions(Pose)
+     */
+    @Override
+    public EntityDimensions getDefaultDimensions(Pose pose) {
+        return this.isBaby() ? BABY_DIMENSIONS : super.getDefaultDimensions(pose);
+    }
+
+    /**
+     * @see Piglin#setBaby(boolean)
+     */
+    @Override
+    public void setBaby(boolean baby) {
+        this.getEntityData().set(DATA_BABY_ID, baby);
+        if (!this.level().isClientSide()) {
+            AttributeInstance speed = this.getAttribute(Attributes.MOVEMENT_SPEED);
+            speed.removeModifier(SPEED_MODIFIER_BABY.id());
+            if (baby) {
+                speed.addTransientModifier(SPEED_MODIFIER_BABY);
+            }
+        }
+    }
+
+    /**
+     * @see Piglin#isBaby()
+     */
+    @Override
+    public boolean isBaby() {
+        return this.getEntityData().get(DATA_BABY_ID);
+    }
+
+    /**
+     * @see Piglin#finalizeSpawn(ServerLevelAccessor, DifficultyInstance, EntitySpawnReason, SpawnGroupData)
+     */
+    @Override
+    public @Nullable SpawnGroupData finalizeSpawn(ServerLevelAccessor level, DifficultyInstance difficulty, EntitySpawnReason spawnReason, @Nullable SpawnGroupData groupData) {
+        RandomSource random = level.getRandom();
+        if (spawnReason != EntitySpawnReason.STRUCTURE) {
+            if (random.nextFloat() < 0.2F) {
+                this.setBaby(true);
+            }
+        }
+
+        return super.finalizeSpawn(level, difficulty, spawnReason, groupData);
+    }
+
+    @Override
+    public boolean removeWhenFarAway(double distanceToClosestPlayer) {
+        return !this.isPersistenceRequired();
+    }
+
+    @Override
+    public boolean canHunt() {
+        return false;
+    }
+
+    @Override
+    public PiglinArmPose getArmPose() {
+        return this.isAggressive() && this.isHoldingMeleeWeapon() ? PiglinArmPose.ATTACKING_WITH_MELEE_WEAPON :
+                PiglinArmPose.DEFAULT;
+    }
+
+    @Nullable
+    @Override
+    public LivingEntity getTarget() {
+        return this.target;
+    }
+
+    @Override
+    public void playAmbientSound() {
+        this.makeSound(this.getAmbientSound());
+    }
+
+    @Override
+    protected SoundEvent getAmbientSound() {
+        return SoundEvents.PIGLIN_AMBIENT;
+    }
+
+    @Override
+    protected SoundEvent getHurtSound(DamageSource damageSource) {
+        return SoundEvents.PIGLIN_HURT;
+    }
+
+    @Override
+    protected SoundEvent getDeathSound() {
+        return SoundEvents.PIGLIN_DEATH;
+    }
+
+    @Override
+    protected void playStepSound(BlockPos pos, BlockState state) {
+        this.playSound(SoundEvents.PIGLIN_STEP, 0.15F, 1.0F);
+    }
+
+    @Override
+    protected void playConvertedSound() {
+        this.makeSound(SoundEvents.PIGLIN_CONVERTED_TO_ZOMBIFIED);
+    }
+}